+++ crates/shell/src/minifb/window_builder.rs
@@ -56,11 +56,7 @@ where
             always_on_top: settings.always_on_top,
             borderless: settings.borderless,
             fonts: settings.fonts,
-            bounds: Rectangle::new(
-                settings.position,
-                settings.size.0,
-                settings.size.1,
-            ),
+            bounds: Rectangle::new(settings.position, settings.size.0, settings.size.1),
             request_receiver: None,
         }
     }