+++ crates/widgets/src/tab_widget.rs
@@ -9,16 +9,15 @@ const BODY_CONTAINER: &str = "body_container";
 The `TabHeaderState` is used to store some callbacks that will be applied during template function.
 Once the template function is called, they are no more used.
 */
-#[derive(Default,AsAny)]
+#[derive(Default, AsAny)]
 pub struct TabHeaderState {
     //Callback called when user click on the header of the tab (normally used to switch to the clicked tab). Only used during initialization
     on_header_click_callback: Option<Box<dyn 'static + Fn(&mut StatesContext, Point) -> bool>>,
     //Callback called when user click on the close button near the tab header (normally used to close the tab). Only used during initialization
-    on_close_click_callback: Option<Box<dyn 'static + Fn(&mut StatesContext, Point) -> bool>>
+    on_close_click_callback: Option<Box<dyn 'static + Fn(&mut StatesContext, Point) -> bool>>,
 }
 
 impl State for TabHeaderState {}
-
 
 widget!(
     /// The `TabHeader` widget is used internally to managed tabs headers. Not meant for other uses.
@@ -74,18 +73,21 @@ widget!(
     }
 );
 
-impl TabHeader
-{
+impl TabHeader {
     ///Set the callback that is called when user click on the header (generally used to switch tab)
-    pub fn on_header_click<T: 'static + Fn(&mut StatesContext, Point) -> bool>(mut self,callback: T)->Self
-    {
+    pub fn on_header_click<T: 'static + Fn(&mut StatesContext, Point) -> bool>(
+        mut self,
+        callback: T,
+    ) -> Self {
         self.state.on_header_click_callback = Some(Box::new(callback));
         self
     }
 
     ///Set the callback that is called when user click on the close button near the header (generally used to close the tab)
-    pub fn on_close_click<T: 'static + Fn(&mut StatesContext, Point) -> bool>(mut self,callback: T)->Self
-    {
+    pub fn on_close_click<T: 'static + Fn(&mut StatesContext, Point) -> bool>(
+        mut self,
+        callback: T,
+    ) -> Self {
         self.state.on_close_click_callback = Some(Box::new(callback));
         self
     }
@@ -94,82 +96,85 @@ impl TabHeader
 impl Template for TabHeader {
     fn template(mut self, id: Entity, ctx: &mut BuildContext) -> Self {
         let mut toggle_button = ToggleButton::new()
-        .background(id)
-        .border_radius(id)
-        .border_width(id)
-        .border_brush(id)
-        .padding(id)
-        .foreground(id)
-        .text(id)
-        .font_size(id)
-        .font(id)
-        .icon(id)
-        .icon_brush(id)
-        .icon_size(id)
-        .icon_font(id)
-        .selected(id)
-        .spacing(id);
+            .background(id)
+            .border_radius(id)
+            .border_width(id)
+            .border_brush(id)
+            .padding(id)
+            .foreground(id)
+            .text(id)
+            .font_size(id)
+            .font(id)
+            .icon(id)
+            .icon_brush(id)
+            .icon_size(id)
+            .icon_font(id)
+            .selected(id)
+            .spacing(id);
 
-        if let Some(callback) = self.state.on_header_click_callback.take() {toggle_button = toggle_button.on_click(callback);}
+        if let Some(callback) = self.state.on_header_click_callback.take() {
+            toggle_button = toggle_button.on_click(callback);
+        }
 
         let mut button = Button::new()
-        .background(id)
-        .border_radius(id)
-        .border_width(id)
-        .border_brush(id)
-        .padding(id)
-        .foreground(id)
-        .font_size(id)
-        .font(id)
-        .icon(id)
-        .icon_brush(id)
-        .icon_size(id)
-        .icon_font(id)
-        .spacing(id)
-        .text("X")
-        .width(2)
-        .visibility(("close_button",id));
+            .background(id)
+            .border_radius(id)
+            .border_width(id)
+            .border_brush(id)
+            .padding(id)
+            .foreground(id)
+            .font_size(id)
+            .font(id)
+            .icon(id)
+            .icon_brush(id)
+            .icon_size(id)
+            .icon_font(id)
+            .spacing(id)
+            .text("X")
+            .width(2)
+            .visibility(("close_button", id));
 
-        if let Some(callback) = self.state.on_close_click_callback.take() {button = button.on_click(callback);}
+        if let Some(callback) = self.state.on_close_click_callback.take() {
+            button = button.on_click(callback);
+        }
         //if self.close_button() == false {button = button.visibility(Visibility::Collapsed);}
 
         self.name("TabWidget")
-        .selected(false)
-        .height(36)
-        .min_width(64)
-        .background(colors::LYNCH_COLOR)
-        .border_radius(4)
-        .border_width(0)
-        .border_brush("transparent")
-        .padding((16, 0, 16, 0))
-        .foreground(colors::LINK_WATER_COLOR)
-        .text("Unnamed Tab")
-        .font_size(fonts::FONT_SIZE_12)
-        .font("Roboto Regular")
-        .icon("")
-        .icon_font("Material Icons")
-        .icon_size(fonts::ICON_FONT_SIZE_12)
-        .icon_brush(colors::LINK_WATER_COLOR)
-        .spacing(8)
-        .close_button(Visibility::Visible)
-        .child(
-            Stack::new()
-            .orientation("horizontal")
-            .child(toggle_button.build(ctx))
-            .child(button.build(ctx))
-            .build(ctx)
-        )
+            .selected(false)
+            .height(36)
+            .min_width(64)
+            .background(colors::LYNCH_COLOR)
+            .border_radius(4)
+            .border_width(0)
+            .border_brush("transparent")
+            .padding((16, 0, 16, 0))
+            .foreground(colors::LINK_WATER_COLOR)
+            .text("Unnamed Tab")
+            .font_size(fonts::FONT_SIZE_12)
+            .font("Roboto Regular")
+            .icon("")
+            .icon_font("Material Icons")
+            .icon_size(fonts::ICON_FONT_SIZE_12)
+            .icon_brush(colors::LINK_WATER_COLOR)
+            .spacing(8)
+            .close_button(Visibility::Visible)
+            .child(
+                Stack::new()
+                    .orientation("horizontal")
+                    .child(toggle_button.build(ctx))
+                    .child(button.build(ctx))
+                    .build(ctx),
+            )
     }
 }
 
 ///Used to internally
-enum TabWidgetAction
-{
+enum TabWidgetAction {
     SelectByIndex(usize),
     SelectByBody(Entity),
-    Add(String,Entity),
+    Add(String, Entity),
     Remove(Entity),
-    SetCloseButtonVisibility(bool)
+    SetCloseButtonVisibility(bool),
 }
 /**
 Through the TabWidgetState it is possible to control the behaviour of TabWidget.
@@ -184,10 +189,10 @@ pub struct TabWidgetState {
     header_container: Entity,
     body_container: Entity,
 
-    tabs: Vec<(Entity,Entity)>,//Header , Body
+    tabs: Vec<(Entity, Entity)>, //Header , Body
     selected: usize,
 
-    close_button_visibility: bool
+    close_button_visibility: bool,
 }
 
 impl TabWidgetState {
@@ -195,16 +200,14 @@ impl TabWidgetState {
     Switch the selected tab with one with the passed index.
     If the index is greater than the tab count, the last one will be selected.
     */
-    pub fn select_by_index(&mut self,index: usize)
-    {
+    pub fn select_by_index(&mut self, index: usize) {
         self.actions.push(TabWidgetAction::SelectByIndex(index));
     }
     /**
     Switch the selected tab with one with the passed body.
     If the no such body is found among the present tabs, nothing happen.
     */
-    pub fn select_by_body(&mut self,entity: Entity)
-    {
+    pub fn select_by_body(&mut self, entity: Entity) {
         self.actions.push(TabWidgetAction::SelectByBody(entity));
     }
 
@@ -212,40 +215,40 @@ impl TabWidgetState {
     Remove the tab with the passed body.
     If the no such body is found among the present tabs, nothing happen.
     */
-    pub fn remove_by_body(&mut self,entity: Entity)
-    {
+    pub fn remove_by_body(&mut self, entity: Entity) {
         self.actions.push(TabWidgetAction::Remove(entity));
     }
 
-
     ///Add a new tab to the widget.
-    pub fn add_tab<T: Into<String>>(&mut self,header: T, body: Entity)
-    {
-        self.actions.push(TabWidgetAction::Add(header.into(),body));
+    pub fn add_tab<T: Into<String>>(&mut self, header: T, body: Entity) {
+        self.actions.push(TabWidgetAction::Add(header.into(), body));
 
         //At this point the tab has not been added yet, so we can check if len is == 0
-        if self.tabs.len() == 0 {self.select_by_index(0);}
+        if self.tabs.len() == 0 {
+            self.select_by_index(0);
+        }
     }
-
 
     ///Set the close button visibility of all the tabs.
-    pub fn set_close_button_visibility(&mut self,value: bool)
-    {
-        self.actions.push(TabWidgetAction::SetCloseButtonVisibility(value));
+    pub fn set_close_button_visibility(&mut self, value: bool) {
+        self.actions
+            .push(TabWidgetAction::SetCloseButtonVisibility(value));
     }
     ///Get the close button visibility status
-    pub fn get_close_button_visibility(&self)->bool{self.close_button_visibility}
+    pub fn get_close_button_visibility(&self) -> bool {
+        self.close_button_visibility
+    }
 
     /**
     Get the tab index associated with the passed body.
     If it is not found, None is returned.
     */
-    pub fn get_index(&self,tab_body: Entity)->Option<usize>
-    {
-        for i in 0..self.tabs.len()
-        {
-            let (_,body) = self.tabs[i];
-            if body == tab_body {return Some(i);}
+    pub fn get_index(&self, tab_body: Entity) -> Option<usize> {
+        for i in 0..self.tabs.len() {
+            let (_, body) = self.tabs[i];
+            if body == tab_body {
+                return Some(i);
+            }
         }
         return None;
     }
@@ -254,103 +257,100 @@ impl TabWidgetState {
     Visually refresh the current tab. This is used when, for example, the selected tab is removed,
     so the tab that take it's place need to be visually updated.
     */
-    fn refresh_selected_tab(&mut self,ctx: &mut Context)
-    {
+    fn refresh_selected_tab(&mut self, ctx: &mut Context) {
         let tab = self.tabs[self.selected];
-        ctx.get_widget(tab.0).set("selected",true);
-        ctx.get_widget(tab.1).set("visibility",Visibility::Visible);
+        ctx.get_widget(tab.0).set("selected", true);
+        ctx.get_widget(tab.1).set("visibility", Visibility::Visible);
     }
 
     /**
     Change the selected tab by index. Unlike the public "select_by_index", this happen immediatly.
     */
-    fn select_by_index_internal(&mut self,ctx: &mut Context,mut index: usize)
-    {
+    fn select_by_index_internal(&mut self, ctx: &mut Context, mut index: usize) {
         //No tabs could be selected if there are no one, so return immediatly
-        if self.tabs.len() == 0 {return;}
+        if self.tabs.len() == 0 {
+            return;
+        }
 
         //If the passed index is greater than tab count, select the last one
-        if index >= self.tabs.len() && index != 0 {index = self.tabs.len()-1;}
+        if index >= self.tabs.len() && index != 0 {
+            index = self.tabs.len() - 1;
+        }
 
-        if self.selected != index
-        {
+        if self.selected != index {
             let current_tab = self.tabs[self.selected];
             let new_tab = self.tabs[index];
 
             //Toggle current button, the new button is toggled by user click
-            ctx.get_widget(current_tab.0).set("selected",false);
+            ctx.get_widget(current_tab.0).set("selected", false);
 
             //Hide current body
-            ctx.get_widget(current_tab.1).set("visibility",Visibility::Hidden);
+            ctx.get_widget(current_tab.1)
+                .set("visibility", Visibility::Hidden);
 
             //The new tab is not "selected" because it is already done by the user click
 
             //Show new body
-            ctx.get_widget(new_tab.1).set("visibility",Visibility::Visible);
+            ctx.get_widget(new_tab.1)
+                .set("visibility", Visibility::Visible);
 
             self.selected = index;
         }
     }
 
-
     ///Add a new tab to the widget. Unlike the public "add_tab", this happen immediatly.
-    fn add_tab_internal(&mut self,ctx: &mut Context,header_text: String,body: Entity)
-    {
+    fn add_tab_internal(&mut self, ctx: &mut Context, header_text: String, body: Entity) {
         //Create the new tab
-        let header = self.create_tab_header(ctx,header_text,body);
+        let header = self.create_tab_header(ctx, header_text, body);
 
         //Set tab body hidden
-        ctx.get_widget(body).set("visibility",Visibility::Hidden);
+        ctx.get_widget(body).set("visibility", Visibility::Hidden);
 
         //Push button to the header container
-        ctx.append_child_entity_to(header,self.header_container);
+        ctx.append_child_entity_to(header, self.header_container);
 
         //Push the body to the body container
-        ctx.append_child_entity_to(body,self.body_container);
+        ctx.append_child_entity_to(body, self.body_container);
 
         //Push the new tab to the list
-        self.tabs.push((header,body));
+        self.tabs.push((header, body));
 
         //If the added tab is the first
-        if self.tabs.len() == 1
-        {
+        if self.tabs.len() == 1 {
             //Select the tab just inserted
             self.selected = 0;
             self.refresh_selected_tab(ctx);
         }
     }
 
-
     ///Remove a tab from the widget. Unlike the public "remove_tab", this happen immediatly.
-    fn remove_tab_internal(&mut self,ctx: &mut Context,body: Entity)
-    {
-        match self.get_index(body)
-        {
-            Some(index)=>
-            {
+    fn remove_tab_internal(&mut self, ctx: &mut Context, body: Entity) {
+        match self.get_index(body) {
+            Some(index) => {
                 //Pop the index tab out of the stored tabs
-                let (header,body) = self.tabs.remove(index);
+                let (header, body) = self.tabs.remove(index);
 
                 //Push button to the header container
-                ctx.remove_child_from(header,self.header_container);
+                ctx.remove_child_from(header, self.header_container);
 
                 //Push the body to the body container
-                ctx.remove_child_from(body,self.body_container);
-
+                ctx.remove_child_from(body, self.body_container);
 
                 //If there is at least one tab
-                if self.tabs.len() != 0
-                {
+                if self.tabs.len() != 0 {
                     //If selected is greater than tab count, select the last one
-                    if self.selected >= self.tabs.len(){self.selected = self.tabs.len()-1;}
+                    if self.selected >= self.tabs.len() {
+                        self.selected = self.tabs.len() - 1;
+                    }
 
                     //Only update current selection if the removed tab is lesser than the selected.
                     //If it is greater, there is no need to update, but simply remove the target tab
-                    if index <= self.selected {self.refresh_selected_tab(ctx);}
+                    if index <= self.selected {
+                        self.refresh_selected_tab(ctx);
+                    }
                 }
-
             }
-            None=>()
+            None => (),
         }
     }
 
@@ -358,59 +358,72 @@ impl TabWidgetState {
     Set the visibility of the close button on all tabs. Unlike the public "set_close_button_visibility", this happen immediatly.
     If the passed "value" is equal to "self.close_button_visibility", so the requested visibility is already present, nothing happen.
     */
-    fn set_close_button_visibility_internal(&mut self,ctx: &mut Context,value: bool)
-    {
-        if self.close_button_visibility != value
-        {
+    fn set_close_button_visibility_internal(&mut self, ctx: &mut Context, value: bool) {
+        if self.close_button_visibility != value {
             self.close_button_visibility = value;
-            let new_visibility = if self.close_button_visibility {Visibility::Visible} else {Visibility::Collapsed};
-            for tab in &self.tabs
-            {
-                ctx.get_widget(tab.0).set("close_button",new_visibility);
+            let new_visibility = if self.close_button_visibility {
+                Visibility::Visible
+            } else {
+                Visibility::Collapsed
+            };
+            for tab in &self.tabs {
+                ctx.get_widget(tab.0).set("close_button", new_visibility);
             }
         }
     }
 
     //Create a new TabHeader entity and return it. For internal use.
-    fn create_tab_header(&self,ctx: &mut Context,text: String,body: Entity)->Entity
-    {
+    fn create_tab_header(&self, ctx: &mut Context, text: String, body: Entity) -> Entity {
         let cloned_entity = ctx.entity;
         TabHeader::new()
-        .visibility(if self.close_button_visibility {Visibility::Visible} else {Visibility::Collapsed})
-        .text(String16::from(text))
-        .on_header_click(move|states,_|{
-            states.get_mut::<TabWidgetState>(cloned_entity).select_by_body(body);
-            true
-        })
-        .on_close_click(move|states,_|{
-            states.get_mut::<TabWidgetState>(cloned_entity).remove_by_body(body);
-            true
-        })
-        .build(&mut ctx.build_context())
+            .visibility(if self.close_button_visibility {
+                Visibility::Visible
+            } else {
+                Visibility::Collapsed
+            })
+            .text(String16::from(text))
+            .on_header_click(move |states, _| {
+                states
+                    .get_mut::<TabWidgetState>(cloned_entity)
+                    .select_by_body(body);
+                true
+            })
+            .on_close_click(move |states, _| {
+                states
+                    .get_mut::<TabWidgetState>(cloned_entity)
+                    .remove_by_body(body);
+                true
+            })
+            .build(&mut ctx.build_context())
     }
 }
 
 impl State for TabWidgetState {
-    fn init(&mut self, _: &mut Registry, ctx: &mut Context)
-    {
+    fn init(&mut self, _: &mut Registry, ctx: &mut Context) {
         self.header_container = ctx.child(HEADER_CONTAINER).entity();
         self.body_container = ctx.child(BODY_CONTAINER).entity();
     }
-    fn update(&mut self, _: &mut Registry, ctx: &mut Context)
-    {
+    fn update(&mut self, _: &mut Registry, ctx: &mut Context) {
         let actions: Vec<TabWidgetAction> = self.actions.drain(..).collect();
-        for action in actions
-        {
-            match action
-            {
-                TabWidgetAction::SelectByIndex(index)=>{self.select_by_index_internal(ctx,index);}
-                TabWidgetAction::SelectByBody(body)=>
-                {
-                    if let Some(index) = self.get_index(body){self.select_by_index_internal(ctx,index)}
+        for action in actions {
+            match action {
+                TabWidgetAction::SelectByIndex(index) => {
+                    self.select_by_index_internal(ctx, index);
                 }
-                TabWidgetAction::Add(header_text,body)=>{self.add_tab_internal(ctx,header_text,body);}
-                TabWidgetAction::Remove(body)=>{self.remove_tab_internal(ctx,body);}
-                TabWidgetAction::SetCloseButtonVisibility(value)=>{self.set_close_button_visibility_internal(ctx,value);}
+                TabWidgetAction::SelectByBody(body) => {
+                    if let Some(index) = self.get_index(body) {
+                        self.select_by_index_internal(ctx, index)
+                    }
+                }
+                TabWidgetAction::Add(header_text, body) => {
+                    self.add_tab_internal(ctx, header_text, body);
+                }
+                TabWidgetAction::Remove(body) => {
+                    self.remove_tab_internal(ctx, body);
+                }
+                TabWidgetAction::SetCloseButtonVisibility(value) => {
+                    self.set_close_button_visibility_internal(ctx, value);
+                }
             }
         }
     }
@@ -438,26 +451,26 @@ widget!(
     }
 );
 
-impl TabWidget
-{
+impl TabWidget {
     ///Set the close button visibility
-    pub fn close_button(mut self,value: bool)->Self
-    {
-        self.state.actions.push(TabWidgetAction::SetCloseButtonVisibility(value));
+    pub fn close_button(mut self, value: bool) -> Self {
+        self.state
+            .actions
+            .push(TabWidgetAction::SetCloseButtonVisibility(value));
         self
     }
     ///Add a tab the widget
-    pub fn tab<T: Into<String>>(mut self,header: T, body: Entity)->Self
-    {
-        self.state.actions.push(TabWidgetAction::Add(header.into(),body));
+    pub fn tab<T: Into<String>>(mut self, header: T, body: Entity) -> Self {
+        self.state
+            .actions
+            .push(TabWidgetAction::Add(header.into(), body));
         self
     }
 }
 
 impl Template for TabWidget {
     fn template(self, id: Entity, ctx: &mut BuildContext) -> Self {
-        self.name("TabWidget")
-        .child(
+        self.name("TabWidget").child(
             Grid::new()
                 .rows(Rows::new().add(32).add("*"))
                 .child(
@@ -465,15 +478,15 @@ impl Template for TabWidget {
                         .id(HEADER_CONTAINER)
                         .orientation("horizontal")
                         .spacing(id)
-                        .build(ctx)
+                        .build(ctx),
                 )
                 .child(
                     Grid::new()
                         .attach(Grid::row(1))
                         .id(BODY_CONTAINER)
-                        .build(ctx)
+                        .build(ctx),
                 )
-                .build(ctx)
+                .build(ctx),
         )
     }
 }