+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Measure distance between two Points
 * Improve: Mouse event arguments
 * Fix: Crash when a child widget is removed
+* TabWidget widget
 
 ### 0.3.1-alpha2